LPS-70229 Remove Base logic for the filter and delegate that in the respective filters
