+++ adaptive-media-asset-publisher-web/build.gradle
@@ -3,7 +3,10 @@ targetCompatibility = "1.8"
 
 dependencies {
 	provided group: "com.liferay", name: "com.liferay.asset.publisher.web", version: "1.0.0"
+	provided group: "com.liferay.portal", name: "com.liferay.portal.kernel", version: "2.7.0"
 	provided group: "javax.portlet", name: "portlet-api", version: "2.0"
+	provided group: "javax.servlet", name: "javax.servlet-api", version: "3.0.1"
 	provided group: "org.osgi", name: "org.osgi.service.component.annotations", version: "1.3.0"
+	provided project(":apps:adaptive-media:adaptive-media-api")
 	provided project(":apps:adaptive-media:adaptive-media-web")
 }
