+++ adaptive-media-blogs-web/src/main/java/com/liferay/adaptive/media/blogs/web/internal/filter/BlogsPortletFilter.java
@@ -14,14 +14,28 @@
 
 package com.liferay.adaptive.media.blogs.web.internal.filter;
 
-import com.liferay.adaptive.media.web.filter.BaseAdaptiveMediaPortletFilter;
+import com.liferay.adaptive.media.processor.content.ContentProcessorHandler;
+import com.liferay.adaptive.media.web.constants.ContentTypes;
 import com.liferay.blogs.web.constants.BlogsPortletKeys;
+import com.liferay.portal.kernel.servlet.BufferCacheServletResponse;
+import com.liferay.portal.kernel.servlet.ServletResponseUtil;
 import com.liferay.portal.kernel.util.ParamUtil;
+import com.liferay.portal.kernel.util.Portal;
 
+import java.io.IOException;
+
+import javax.portlet.PortletException;
 import javax.portlet.RenderRequest;
+import javax.portlet.RenderResponse;
+import javax.portlet.filter.FilterChain;
+import javax.portlet.filter.FilterConfig;
 import javax.portlet.filter.PortletFilter;
+import javax.portlet.filter.RenderFilter;
+
+import javax.servlet.http.HttpServletResponse;
 
 import org.osgi.service.component.annotations.Component;
+import org.osgi.service.component.annotations.Reference;
 
 /**
  * @author Alejandro Tardín
@@ -31,14 +45,49 @@ import org.osgi.service.component.annotations.Component;
 	property = {"javax.portlet.name=" + BlogsPortletKeys.BLOGS},
 	service = PortletFilter.class
 )
-public class BlogsPortletFilter extends BaseAdaptiveMediaPortletFilter {
+public class BlogsPortletFilter implements RenderFilter {
 
 	@Override
-	protected boolean mustProcessContent(RenderRequest renderRequest) {
+	public void destroy() {
+	}
+
+	@Override
+	public void doFilter(
+			RenderRequest renderRequest, RenderResponse renderResponse,
+			FilterChain filterChain)
+		throws IOException, PortletException {
+
+		filterChain.doFilter(renderRequest, renderResponse);
+
 		String mvcRenderCommandName = ParamUtil.getString(
 			renderRequest, "mvcRenderCommandName");
 
-		return "/blogs/view_entry".equals(mvcRenderCommandName);
+		if (!mvcRenderCommandName.equals("/blogs/view_entry")) {
+			return;
+		}
+
+		HttpServletResponse httpServletResponse = portal.getHttpServletResponse(
+			renderResponse);
+
+		BufferCacheServletResponse bufferCacheServletResponse =
+			(BufferCacheServletResponse)httpServletResponse;
+
+		String content = bufferCacheServletResponse.getString();
+
+		String processedContent = contentProcessorHandler.process(
+			ContentTypes.HTML, content);
+
+		ServletResponseUtil.write(httpServletResponse, processedContent);
 	}
+
+	@Override
+	public void init(FilterConfig filterConfig) {
+	}
+
+	@Reference
+	protected ContentProcessorHandler contentProcessorHandler;
+
+	@Reference
+	protected Portal portal;
 
 }
