+++ adaptive-media-web/bnd.bnd
@@ -1,9 +1,7 @@
 Bundle-Name: Liferay Adaptive Media Web
 Bundle-SymbolicName: com.liferay.adaptive.media.web
 Bundle-Version: 1.0.0
-Export-Package:\
-	com.liferay.adaptive.media.web.constants,\
-	com.liferay.adaptive.media.web.filter
+Export-Package: com.liferay.adaptive.media.web.constants
 Liferay-Releng-Module-Group-Description:
 Liferay-Releng-Module-Group-Title: Adaptive Media
 Web-ContextPath: /adaptive-media-web
