--- adaptive-media-web/src/main/java/com/liferay/adaptive/media/web/filter/BaseAdaptiveMediaPortletFilter.java
@@ -1,81 +1,0 @@
-/**
- * Copyright (c) 2000-present Liferay, Inc. All rights reserved.
- *
- * This library is free software; you can redistribute it and/or modify it under
- * the terms of the GNU Lesser General Public License as published by the Free
- * Software Foundation; either version 2.1 of the License, or (at your option)
- * any later version.
- *
- * This library is distributed in the hope that it will be useful, but WITHOUT
- * ANY WARRANTY; without even the implied warranty of MERCHANTABILITY or FITNESS
- * FOR A PARTICULAR PURPOSE. See the GNU Lesser General Public License for more
- * details.
- */
-
-package com.liferay.adaptive.media.web.filter;
-
-import com.liferay.adaptive.media.processor.content.ContentProcessorHandler;
-import com.liferay.adaptive.media.web.constants.ContentTypes;
-import com.liferay.portal.kernel.servlet.BufferCacheServletResponse;
-import com.liferay.portal.kernel.servlet.ServletResponseUtil;
-import com.liferay.portal.kernel.util.Portal;
-
-import java.io.IOException;
-
-import javax.portlet.PortletException;
-import javax.portlet.RenderRequest;
-import javax.portlet.RenderResponse;
-import javax.portlet.filter.FilterChain;
-import javax.portlet.filter.FilterConfig;
-import javax.portlet.filter.RenderFilter;
-
-import javax.servlet.http.HttpServletResponse;
-
-import org.osgi.service.component.annotations.Reference;
-
-/**
- * @author Alejandro Tardín
- */
-public abstract class BaseAdaptiveMediaPortletFilter implements RenderFilter {
-
-	@Override
-	public void destroy() {
-	}
-
-	@Override
-	public void doFilter(
-			RenderRequest renderRequest, RenderResponse renderResponse,
-			FilterChain filterChain)
-		throws IOException, PortletException {
-
-		filterChain.doFilter(renderRequest, renderResponse);
-
-		if (mustProcessContent(renderRequest)) {
-			HttpServletResponse httpServletResponse =
-				portal.getHttpServletResponse(renderResponse);
-
-			BufferCacheServletResponse bufferCacheServletResponse =
-				(BufferCacheServletResponse)httpServletResponse;
-
-			String content = bufferCacheServletResponse.getString();
-
-			String processedContent = contentProcessorHandler.process(
-				ContentTypes.HTML, content);
-
-			ServletResponseUtil.write(httpServletResponse, processedContent);
-		}
-	}
-
-	@Override
-	public void init(FilterConfig filterConfig) {
-	}
-
-	protected abstract boolean mustProcessContent(RenderRequest renderRequest);
-
-	@Reference
-	protected ContentProcessorHandler contentProcessorHandler;
-
-	@Reference
-	protected Portal portal;
-
-}
